+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### New Features
 
+* Added support for assignments within expressions (e.g., `x = ++y;`)
 * Added support for excluding the conversion of unbased unsized literals (e.g.,
   `'1`, `'x`) via `--exclude UnbasedUniszed`
 * Added support for enumerated type ranges (e.g., `enum { X[3:5] }`)